requirements.txt
streamlit==1.28.1
pandas==2.1.1
pymongo==4.5.0
plotly==5.17.0
python-dotenv==1.0.0